Unify PPOAgent and RNDPPOAgent
Allow for better diff in GitHub
